better background on iOS
iOS safari doesn't implement fixed backgrounds, probably for
battery saving reasons, and the bg image was streched on iPhones.

As a workaround I'm not fixing the image on smaller screens and
letting it repeat. The effect is not as nice but the image is not
as important on small screens anyways.