switch from astro's aplinejs plugin to manual alpinejs
I run into issues in Table.astro, the inline script element did not work.
This is still not ideal, the component initializes Alpine, if I ever
have multiple Table components on one page they will conflict but it
should be fine for now.

This way I also get typescript support for the alpine code and
Alpine will be initialized where it is needed.